updated nugets + seperate AboutModel
