+++ Aleph1.Skeletons.WebAPI/Aleph1.Skeletons.WebAPI.BL.Implementation/packages.config
@@ -1,9 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Aleph1.DI.Contracts" version="3.0.2" targetFramework="net471" />
-  <package id="Aleph1.Logging" version="5.0.1" targetFramework="net471" />
-  <package id="Aleph1.Utilities" version="3.1.4" targetFramework="net471" />
+  <package id="Aleph1.DI.Contracts" version="3.3.0" targetFramework="net471" />
+  <package id="Aleph1.Logging" version="5.1.0" targetFramework="net471" />
+  <package id="Aleph1.Utilities" version="3.3.0" targetFramework="net471" />
   <package id="NLog" version="4.5.11" targetFramework="net471" />
-  <package id="PostSharp" version="6.0.32" targetFramework="net471" developmentDependency="true" />
-  <package id="PostSharp.Redist" version="6.0.32" targetFramework="net471" />
+  <package id="PostSharp" version="6.0.34" targetFramework="net471" developmentDependency="true" />
+  <package id="PostSharp.Redist" version="6.0.34" targetFramework="net471" />
 </packages>
