+++ Aleph1.Skeletons.WebAPI/Aleph1.Skeletons.WebAPI.DAL.Moq/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Aleph1.DI.Contracts" version="3.0.2" targetFramework="net471" />
+  <package id="Aleph1.DI.Contracts" version="3.3.0" targetFramework="net471" />
 </packages>
