+++ Aleph1.Skeletons.WebAPI/Aleph1.Skeletons.WebAPI.Security.Implementation/Aleph1.Skeletons.WebAPI.Security.Implementation.csproj
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="..\packages\PostSharp.6.0.32\build\PostSharp.props" Condition="Exists('..\packages\PostSharp.6.0.32\build\PostSharp.props')" />
+  <Import Project="..\packages\PostSharp.6.0.34\build\PostSharp.props" Condition="Exists('..\packages\PostSharp.6.0.34\build\PostSharp.props')" />
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Dev</Configuration>
@@ -40,20 +40,20 @@
     <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Aleph1.DI.Contracts, Version=3.0.2.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Aleph1.DI.Contracts.3.0.2\lib\net471\Aleph1.DI.Contracts.dll</HintPath>
+    <Reference Include="Aleph1.DI.Contracts, Version=3.3.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\Aleph1.DI.Contracts.3.3.0\lib\net471\Aleph1.DI.Contracts.dll</HintPath>
     </Reference>
-    <Reference Include="Aleph1.Logging, Version=5.0.1.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Aleph1.Logging.5.0.1\lib\net471\Aleph1.Logging.dll</HintPath>
+    <Reference Include="Aleph1.Logging, Version=5.1.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\Aleph1.Logging.5.1.0\lib\net471\Aleph1.Logging.dll</HintPath>
     </Reference>
-    <Reference Include="Aleph1.Security.Contracts, Version=3.1.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Aleph1.Security.Contracts.3.1.0\lib\net471\Aleph1.Security.Contracts.dll</HintPath>
+    <Reference Include="Aleph1.Security.Contracts, Version=3.3.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\Aleph1.Security.Contracts.3.3.0\lib\net471\Aleph1.Security.Contracts.dll</HintPath>
     </Reference>
-    <Reference Include="Aleph1.Security.Implementation.3DES, Version=3.1.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Aleph1.Security.Implementation.3DES.3.1.0\lib\net471\Aleph1.Security.Implementation.3DES.dll</HintPath>
+    <Reference Include="Aleph1.Security.Implementation.3DES, Version=3.3.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\Aleph1.Security.Implementation.3DES.3.3.0\lib\net471\Aleph1.Security.Implementation.3DES.dll</HintPath>
     </Reference>
-    <Reference Include="Aleph1.Utilities, Version=3.1.4.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Aleph1.Utilities.3.1.4\lib\net471\Aleph1.Utilities.dll</HintPath>
+    <Reference Include="Aleph1.Utilities, Version=3.3.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\Aleph1.Utilities.3.3.0\lib\net471\Aleph1.Utilities.dll</HintPath>
     </Reference>
     <Reference Include="Newtonsoft.Json, Version=12.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <HintPath>..\packages\Newtonsoft.Json.12.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
@@ -61,8 +61,8 @@
     <Reference Include="NLog, Version=4.0.0.0, Culture=neutral, PublicKeyToken=5120e14c03d0593c, processorArchitecture=MSIL">
       <HintPath>..\packages\NLog.4.5.11\lib\net45\NLog.dll</HintPath>
     </Reference>
-    <Reference Include="PostSharp, Version=6.0.32.0, Culture=neutral, PublicKeyToken=b13fd38b8f9c99d7, processorArchitecture=MSIL">
-      <HintPath>..\packages\PostSharp.Redist.6.0.32\lib\net45\PostSharp.dll</HintPath>
+    <Reference Include="PostSharp, Version=6.0.34.0, Culture=neutral, PublicKeyToken=b13fd38b8f9c99d7, processorArchitecture=MSIL">
+      <HintPath>..\packages\PostSharp.Redist.6.0.34\lib\net45\PostSharp.dll</HintPath>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.ComponentModel.Composition" />
@@ -107,8 +107,8 @@
     <PropertyGroup>
       <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
     </PropertyGroup>
-    <Error Condition="!Exists('..\packages\PostSharp.6.0.32\build\PostSharp.props')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\PostSharp.6.0.32\build\PostSharp.props'))" />
-    <Error Condition="!Exists('..\packages\PostSharp.6.0.32\build\PostSharp.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\PostSharp.6.0.32\build\PostSharp.targets'))" />
+    <Error Condition="!Exists('..\packages\PostSharp.6.0.34\build\PostSharp.props')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\PostSharp.6.0.34\build\PostSharp.props'))" />
+    <Error Condition="!Exists('..\packages\PostSharp.6.0.34\build\PostSharp.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\PostSharp.6.0.34\build\PostSharp.targets'))" />
   </Target>
-  <Import Project="..\packages\PostSharp.6.0.32\build\PostSharp.targets" Condition="Exists('..\packages\PostSharp.6.0.32\build\PostSharp.targets')" />
+  <Import Project="..\packages\PostSharp.6.0.34\build\PostSharp.targets" Condition="Exists('..\packages\PostSharp.6.0.34\build\PostSharp.targets')" />
 </Project>
