+++ Aleph1.Skeletons.WebAPI/Aleph1.Skeletons.WebAPI.WebAPI/Controllers/AboutController.cs
@@ -1,5 +1,6 @@
 ﻿using Aleph1.Logging;
 using Aleph1.Skeletons.WebAPI.WebAPI.Classes;
+using Aleph1.Skeletons.WebAPI.WebAPI.Models;
 using Aleph1.WebAPI.ExceptionHandler;
 using System;
 using System.Reflection;
@@ -26,25 +27,6 @@ namespace Aleph1.Skeletons.WebAPI.WebAPI.Controllers
                 APIVersion = Assembly.GetExecutingAssembly()?.GetName()?.Version?.ToString(),
                 Server = Environment.MachineName
             };
-        }
-
-        /// <summary>Some data about the current api and user</summary>
-        public class AboutModel
-        {
-            /// <summary>client ip.</summary>
-            public string ClientIP { get; set; }
-
-            /// <summary>logon name of the client user.</summary>
-            public string ClientUserName { get; set; }
-
-            /// <summary>API version. </summary>
-            public string APIVersion { get; set; }
-
-            /// <summary>the environment (DEV/TEST/PROD)</summary>
-            public string Environment { get; set; }
-
-            /// <summary>server name.</summary>
-            public string Server { get; set; }
         }
     }
 }
