+++ Aleph1.Skeletons.WebAPI/Aleph1.Skeletons.WebAPI.WebAPI/index.html
@@ -37,7 +37,7 @@
         <h1>
             Aleph1.Skeletons.WebAPI.WebAPI -
             <a href="https://github.com/avrahamcool/Aleph1.Skeletons/blob/master/CHANGELOG.md">
-                Aleph1.Skeletons 2.2.0
+                Aleph1.Skeletons 2.4.0
             </a>
         </h1>
     </div>
