+++ Aleph1.Skeletons.WebAPI/Aleph1.Skeletons.WebAPI.WebAPI/packages.config
@@ -1,37 +1,38 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Aleph1.DI.Contracts" version="3.0.2" targetFramework="net471" />
-  <package id="Aleph1.DI.CustomConfigurationSection" version="3.0.1" targetFramework="net471" />
-  <package id="Aleph1.DI.UnityImplementation" version="3.0.5" targetFramework="net471" />
-  <package id="Aleph1.Logging" version="5.0.1" targetFramework="net471" />
-  <package id="Aleph1.Utilities" version="3.1.4" targetFramework="net471" />
-  <package id="Aleph1.WebAPI.ExceptionHandler" version="3.0.2" targetFramework="net471" />
-  <package id="FluentValidation" version="8.1.2" targetFramework="net471" />
-  <package id="FluentValidation.ValidatorAttribute" version="8.1.2" targetFramework="net471" />
-  <package id="FluentValidation.WebApi" version="8.1.2" targetFramework="net471" />
+  <package id="Aleph1.DI.Contracts" version="3.3.0" targetFramework="net471" />
+  <package id="Aleph1.DI.CustomConfigurationSection" version="3.3.0" targetFramework="net471" />
+  <package id="Aleph1.DI.UnityImplementation" version="3.3.0" targetFramework="net471" />
+  <package id="Aleph1.Logging" version="5.1.0" targetFramework="net471" />
+  <package id="Aleph1.Utilities" version="3.3.0" targetFramework="net471" />
+  <package id="Aleph1.WebAPI.ExceptionHandler" version="3.1.0" targetFramework="net471" />
+  <package id="FluentValidation" version="8.1.3" targetFramework="net471" />
+  <package id="FluentValidation.ValidatorAttribute" version="8.1.3" targetFramework="net471" />
+  <package id="FluentValidation.WebApi" version="8.1.3" targetFramework="net471" />
   <package id="Microsoft.AspNet.WebApi" version="5.2.7" targetFramework="net471" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net471" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net471" />
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.7" targetFramework="net471" />
   <package id="Microsoft.Net.Compilers" version="2.10.0" targetFramework="net471" developmentDependency="true" />
-  <package id="Microsoft.Owin" version="4.0.0" targetFramework="net471" />
+  <package id="Microsoft.Owin" version="4.0.1" targetFramework="net471" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net471" />
   <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net471" />
   <package id="NLog" version="4.5.11" targetFramework="net471" />
   <package id="NLog.Schema" version="4.5.11" targetFramework="net471" />
-  <package id="NLog.Web" version="4.7.0" targetFramework="net471" />
+  <package id="NLog.Web" version="4.8.0" targetFramework="net471" />
   <package id="Owin" version="1.0" targetFramework="net471" />
-  <package id="PostSharp" version="6.0.32" targetFramework="net471" developmentDependency="true" />
-  <package id="PostSharp.Redist" version="6.0.32" targetFramework="net471" />
+  <package id="PostSharp" version="6.0.34" targetFramework="net471" developmentDependency="true" />
+  <package id="PostSharp.Redist" version="6.0.34" targetFramework="net471" />
   <package id="Swashbuckle" version="5.6.0" targetFramework="net471" />
   <package id="Swashbuckle.Core" version="5.6.0" targetFramework="net471" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net471" />
   <package id="System.ComponentModel.Annotations" version="4.5.0" targetFramework="net471" />
   <package id="System.ComponentModel.Primitives" version="4.3.0" targetFramework="net471" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net471" />
   <package id="System.ValueTuple" version="4.5.0" targetFramework="net471" />
-  <package id="Unity.Abstractions" version="3.3.1" targetFramework="net471" />
-  <package id="Unity.AspNet.WebApi" version="5.0.15" targetFramework="net471" />
-  <package id="Unity.Container" version="5.8.13" targetFramework="net471" />
+  <package id="Unity.Abstractions" version="4.0.2" targetFramework="net471" />
+  <package id="Unity.AspNet.WebApi" version="5.9.1" targetFramework="net471" />
+  <package id="Unity.Container" version="5.9.3" targetFramework="net471" />
   <package id="WebActivatorEx" version="2.2.0" targetFramework="net471" />
   <package id="WebApiThrottle" version="1.5.4" targetFramework="net471" />
 </packages>
